+++ packages/dtslint/src/util.ts
@@ -1,4 +1,3 @@
-import assert = require("assert");
 import fs from "fs";
 import { basename, dirname, join } from "path";
 import stripJsonComments = require("strip-json-comments");
@@ -25,7 +24,23 @@ export function getCompilerOptions(dirPath: string): ts.CompilerOptions {
   return readJson(tsconfigPath).compilerOptions as ts.CompilerOptions;
 }
 
-export function last<T>(a: readonly T[]): T {
-  assert(a.length !== 0);
-  return a[a.length - 1];
+export function isMainFile(fileName: string, allowNested: boolean) {
+  // Linter may be run with cwd of the package. We want `index.d.ts` but not `submodule/index.d.ts` to match.
+  if (fileName === "index.d.ts") {
+    return true;
+  }
+
+  if (basename(fileName) !== "index.d.ts") {
+    return false;
+  }
+
+  let parent = dirname(fileName);
+  // May be a directory for an older version, e.g. `v0`.
+  // Note a types redirect `foo/ts3.1` should not have its own header.
+  if (allowNested && /^v(0\.)?\d+$/.test(basename(parent))) {
+    parent = dirname(parent);
+  }
+
+  // Allow "types/foo/index.d.ts", not "types/foo/utils/index.d.ts"
+  return basename(dirname(parent)) === "types";
 }