+++ packages/eslint-plugin/src/rules/index.ts
@@ -17,6 +17,7 @@ import noSingleDeclareModule = require("./no-single-declare-module");
 import noOldDTHeader = require("./no-old-dt-header");
 import noImportOfDevDependencies = require("./no-import-of-dev-dependencies");
 import npmNaming = require("./npm-naming");
+import expect = require("./expect");
 
 export const rules = {
   "export-just-namespace": exportJustNamespace,
@@ -38,4 +39,5 @@ export const rules = {
   "no-old-dt-header": noOldDTHeader,
   "no-import-of-dev-dependencies": noImportOfDevDependencies,
   "npm-naming": npmNaming,
+  expect,
 };